Move HtmlPrinter and HtmlPrettyPrinter into tests
